@@ -1,19 +1,21 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
-  <PropertyGroup>
-    <TargetFrameworks>netcoreapp3.1;net5.0;</TargetFrameworks>
-    <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
-    <PackageId>ppioli.Telegram.Sink</PackageId>
-    <Authors>Pablo Pioli</Authors>
-    <Description>A Telegram sink for Serilog</Description>
-    <PackageLicenseExpression>MIT</PackageLicenseExpression>
-    <RepositoryUrl>https://github.com/pablopioli/telegram-sink</RepositoryUrl>
-    <PackageProjectUrl>https://github.com/pablopioli/telegram-sink</PackageProjectUrl>
-    <Version>1.1.0</Version>
-  </PropertyGroup>
+	<PropertyGroup>
+		<TargetFrameworks>net6.0;</TargetFrameworks>
+		<GeneratePackageOnBuild>true</GeneratePackageOnBuild>
+		<PackageId>ppioli.Telegram.Sink</PackageId>
+		<Authors>Pablo Pioli</Authors>
+		<Description>A Telegram sink for Serilog</Description>
+		<PackageLicenseExpression>MIT</PackageLicenseExpression>
+		<RepositoryUrl>https://github.com/pablopioli/telegram-sink</RepositoryUrl>
+		<PackageProjectUrl>https://github.com/pablopioli/telegram-sink</PackageProjectUrl>
+		<Version>1.2.0</Version>
+		<ImplicitUsings>enable</ImplicitUsings>
+		<Nullable>enable</Nullable>
+	</PropertyGroup>
 
-  <ItemGroup>
-    <PackageReference Include="Serilog" Version="2.10.0" />
-  </ItemGroup>
+	<ItemGroup>
+		<PackageReference Include="Serilog" Version="2.10.0" />
+	</ItemGroup>
 
 </Project>
